pre-commit run --all

+++ src/briefcase/protocols/v2-periphery/libraries/UniswapV2OracleLibrary.sol
@@ -1,7 +1,7 @@
 pragma solidity >=0.5.0;
 
-import {FixedPoint} from "../../lib-external/solidity-lib/contracts/libraries/FixedPoint.sol";
-import {IUniswapV2Pair} from "../../v2-core/interfaces/IUniswapV2Pair.sol";
+import {FixedPoint} from '../../lib-external/solidity-lib/contracts/libraries/FixedPoint.sol';
+import {IUniswapV2Pair} from '../../v2-core/interfaces/IUniswapV2Pair.sol';
 
 // library with helper methods for oracles that are concerned with computing average prices
 library UniswapV2OracleLibrary {
@@ -13,9 +13,11 @@ library UniswapV2OracleLibrary {
     }
 
     // produces the cumulative price using counterfactuals to save gas and avoid a call to sync.
-    function currentCumulativePrices(
-        address pair
-    ) internal view returns (uint price0Cumulative, uint price1Cumulative, uint32 blockTimestamp) {
+    function currentCumulativePrices(address pair)
+        internal
+        view
+        returns (uint256 price0Cumulative, uint256 price1Cumulative, uint32 blockTimestamp)
+    {
         blockTimestamp = currentBlockTimestamp();
         price0Cumulative = IUniswapV2Pair(pair).price0CumulativeLast();
         price1Cumulative = IUniswapV2Pair(pair).price1CumulativeLast();
@@ -27,9 +29,9 @@ library UniswapV2OracleLibrary {
             uint32 timeElapsed = blockTimestamp - blockTimestampLast;
             // addition overflow is desired
             // counterfactual
-            price0Cumulative += uint(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
+            price0Cumulative += uint256(FixedPoint.fraction(reserve1, reserve0)._x) * timeElapsed;
             // counterfactual
-            price1Cumulative += uint(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
+            price1Cumulative += uint256(FixedPoint.fraction(reserve0, reserve1)._x) * timeElapsed;
         }
     }
 }
